Tab-ify supporting points, comments, history.
* I rolled the history template into the view
point template.

* I left the route to the point history, the 
handler and the template intact.  Maybe we want 
to use it later?

* Also, Aaron you probably want to look at and 
clean up the view point handler, I rather crudely 
rolled in the data needed for the history view 
into it. I'm not sure how the point model tree 
works so the way I'm pulling that data is probably 
redundant and/or inefficient.