Point scoring update to make the upvote/downvote more consistent:
max(up - down + 1, 0)

Previously, with the ((up - down) or 1), there was no change for the first upvote, and 1 up - 1 down -> 0, which would default back to 1.

The abs value was removed from the second term since +1 and -1 net should be treated separately.  And max(0 was added so continual downvotes to a subpoint don't affect the score.

pointContent: Added score to stats for debugging/display.