Get relevance rating working
I had to introduce some slightly funky stuff here - there's now a Link
object in the graphql schema that we create slightly differently in
different circumstances. We store a bit of extra information on
supporting/counter points to enable giving Links consistent IDs that
make apollo's denormalization work for us.

The basic shape of this will stay the same even after we overhaul the
bakcend, but the implementation of this definitely shows how creaky
the backend is!