Initial work to lazily load subpoint data
We shouldn't be loading subpoints until users indicate they might want
to see them. In this PoC, load subpoint data when users click on
"expand." In the future we could load data when they hover over
"expand" for less lag.

This definitely breaks subpoint expansion - we need to figure out how
to make this work properly