First pass at existing claim adder
Implements:

```
Option 2: Unified Text Field
I think this is where we want to get to eventually. Feels fluid and elegant to me. Dunno how hard it is?

Clicking Add Evidence brings up a text input field with placeholder text along the lines of “Make a claim or Search for an existing claim”
Clicking inside the text input field brings up, below it, the Recently Viewed list
If a user clicks on a Recently Viewed claim, it gets added
If user starts to type, the Recently Viewed list is replaced by suggestions via Search.
If searching on every keystroke is costly, we could search only on spacebar hits
If user doesn’t see a suggestion s/he likes, user ends up typing in a new claim and clicks “Add”.
```