Get unified title text feedback area working
We're moving toward a unified metaphor for dealing with claim titles -
we want to show feedback about new claim titles in the same dropdown
where we let people pick existing claims to add as evidence.

This commit gets the errors displaying in the same place, and also
moves responsibility for filtering suggestions to a new component that
is very similar to the ClaimSearch component (which is kept around
even though it is no longer used because we will likely want it in the
future).